+++ bgpd/bgp_vty.h
@@ -72,6 +72,5 @@ extern int bgp_vty_find_and_parse_afi_safi_bgp(struct vty *vty,
 					       bool use_json);
 extern int bgp_show_summary_vty(struct vty *vty, const char *name, afi_t afi,
 				safi_t safi, bool show_failed, bool use_json);
-extern void bgp_vpn_policy_config_write_afi(struct vty *vty, struct bgp *bgp,
-					    afi_t afi);
+
 #endif /* _QUAGGA_BGP_VTY_H */
